+++ framework/py/flwr/server/run_serverapp.py
@@ -20,8 +20,8 @@ from logging import DEBUG
 from flwr.common import Context
 from flwr.common.logger import log
 from flwr.common.object_ref import load_app
+from flwr.serverapp.grid import Grid
 
-from .grid import Grid
 from .server_app import LoadServerAppError, ServerApp
